Add files via upload
Here is a python program created that is a text based hangman game where the program initially asks the user to enter their name, greets them and tells how the game works where the user will have to guess a random letter to unlock the word and win the game . The user will be given 5 chances to guess the letter and unlock the word , incase the user is not able to unlock the word with the five chances the game ends by displaying a message telling "you lost" and displaying the word given to the user to guess at the end . If the user enter the letter that he/she has already guessed , then the program will display a message telling "you have already guessed the letter". If the user succeeds in guessing the word the game ends by displaying the message "you win".